requirements.txt
Werkzeug==2.0.1
tmdbsimple==2.8.0
Flask==2.0.1
PyYAML==5.4.1
